refactor: inline createReflectionFlowContext into runReflectionFlow
Collapse one abstraction layer by moving service creation directly
into the flow runner:

- Delete ReflectionFlowContext.ts (221 lines)
- Inline service creation in runReflectionFlow
- Remove intermediate context object and types
- Create simple interruptible object for registration

Net reduction: ~163 lines. Services are now created directly where
they're used, eliminating the factory indirection.
